refactor: 컴포넌트 경로 수정 및 UI 구성 요소 통합, 불필요한 파일 삭제, 새로운 위젯 추가
