+++ src/app/signup/page.tsx
@@ -1,4 +1,4 @@
-import { SignupView } from "@/views/SignupView";
+import { SignupView } from "@/views";
 
 export default function SignupPage() {
   return <SignupView />;
