+++ src/entities/menu/ui/MenuCard.tsx
@@ -1,10 +1,9 @@
 "use client";
 
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { cn } from "@/shared/lib/utils";
-import { MenuItemType } from "@/shared/types";
+import { cn, MenuItemType } from "@/shared";
+import { Badge } from "@/shared/ui/badge";
+import { Button } from "@/shared/ui/button";
+import { Card, CardContent, CardHeader, CardTitle } from "@/shared/ui/card";
 import { Edit } from "lucide-react";
 
 interface MenuCardProps {
