--- src/entities/user/index.ts
@@ -1,2 +1,0 @@
-export * from "./model/user-api";
-export * from "./types";
