--- src/features/auth/index.ts
@@ -1,2 +1,0 @@
-export * from "./signin";
-export * from "./signup";
