--- src/features/auth/signin/index.ts
@@ -1,2 +1,0 @@
-export { signinSchema } from "./lib/validation";
-export { SigninForm } from "./ui/signin-form";
