--- src/features/auth/signup/index.ts
@@ -1,2 +1,0 @@
-export { signupSchema } from "./lib/validation";
-export { SignupForm } from "./ui/signup-form";
