+++ src/features/menu-manage/add-menu-dialog/ui/AddMenuDialog.tsx
@@ -1,18 +1,18 @@
 "use client";
-import { Button } from "@/components/ui/button";
+import { useCreateMenu } from "@/entities";
+import { CreateMenuRequestType } from "@/shared";
+import { Button } from "@/shared/ui/button";
 import {
   Dialog,
   DialogContent,
   DialogHeader,
   DialogTitle,
   DialogTrigger,
-} from "@/components/ui/dialog";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Switch } from "@/components/ui/switch";
-import { Textarea } from "@/components/ui/textarea";
-import { useCreateMenu } from "@/entities";
-import { CreateMenuRequestType } from "@/shared/types";
+} from "@/shared/ui/dialog";
+import { Input } from "@/shared/ui/input";
+import { Label } from "@/shared/ui/label";
+import { Switch } from "@/shared/ui/switch";
+import { Textarea } from "@/shared/ui/textarea";
 import { Plus } from "lucide-react";
 import { useState } from "react";
