+++ src/provider/index.ts
@@ -1,0 +1,1 @@
+export { ReactQueryClientProvider } from "./react-query-client-provider/ReactQueryClientProvider";
