+++ src/shared/api/index.ts
@@ -1,0 +1,3 @@
+export { menuAPI } from "./menu";
+export { orderAPI } from "./order";
+export { userAPI } from "./user";
