+++ src/views/SalesManageView/SalesManageView.tsx
