+++ src/widgets/index.ts
@@ -3,4 +3,5 @@ export { ConfirmedOrderList } from "./confirmed-order-list/ui/ConfirmedOrderList
 export { InprogressOrderList } from "./inprogress-order-list/ui/InprogressOrderList";
 export { MenuCart } from "./menu-cart/ui/MenuCart";
 export { MenuList } from "./menu-list/ui/MenuList";
+export { NavigationBar } from "./navigation-bar/ui/NavigationBar";
 export { OrderPieChart } from "./order-piechart/ui/OrderPieChart";
